Reorganize molgenis-data-index package structure.
