+++ molgenis-data-elasticsearch/src/main/java/org/molgenis/data/elasticsearch/client/ElasticsearchConfig.java
@@ -1,7 +1,7 @@
 package org.molgenis.data.elasticsearch.client;
 
 import org.elasticsearch.client.Client;
-import org.molgenis.data.elasticsearch.index.IndexConfig;
+import org.molgenis.data.index.IndexConfig;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
