+++ molgenis-data-index/src/main/java/org/molgenis/data/index/bootstrap/IndexBootstrapper.java
@@ -1,10 +1,10 @@
-package org.molgenis.data.elasticsearch.bootstrap;
+package org.molgenis.data.index.bootstrap;
 
 import org.molgenis.data.DataService;
-import org.molgenis.data.elasticsearch.index.job.IndexJobExecution;
-import org.molgenis.data.elasticsearch.index.job.IndexJobExecutionMeta;
 import org.molgenis.data.index.IndexActionRegisterService;
 import org.molgenis.data.index.IndexService;
+import org.molgenis.data.index.job.IndexJobExecution;
+import org.molgenis.data.index.job.IndexJobExecutionMeta;
 import org.molgenis.data.index.meta.IndexAction;
 import org.molgenis.data.index.meta.IndexActionMetaData;
 import org.molgenis.data.jobs.model.JobExecutionMetaData;
