+++ molgenis-data-index/src/test/java/org/molgenis/data/index/job/IndexJobServiceTest.java
@@ -1,4 +1,4 @@
-package org.molgenis.data.elasticsearch.index.job;
+package org.molgenis.data.index.job;
 
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
