+++ molgenis-platform-integration-tests/src/test/java/org/molgenis/integrationtest/platform/OneToManyIT.java
@@ -5,7 +5,7 @@ import com.google.common.collect.Lists;
 import org.molgenis.data.DataService;
 import org.molgenis.data.Entity;
 import org.molgenis.data.OneToManyTestHarness;
-import org.molgenis.data.elasticsearch.index.job.IndexJobScheduler;
+import org.molgenis.data.index.job.IndexJobScheduler;
 import org.molgenis.data.staticentity.bidirectional.authorbook1.AuthorMetaData1;
 import org.molgenis.data.staticentity.bidirectional.authorbook1.BookMetaData1;
 import org.molgenis.data.staticentity.bidirectional.person1.PersonMetaData1;
