+++ molgenis-platform-integration-tests/src/test/java/org/molgenis/integrationtest/platform/datatypeediting/AbstractAttributeTypeUpdateIT.java
@@ -4,7 +4,7 @@ import org.molgenis.auth.UserAuthorityMetaData;
 import org.molgenis.data.DataService;
 import org.molgenis.data.Entity;
 import org.molgenis.data.EntityManager;
-import org.molgenis.data.elasticsearch.index.job.IndexJobScheduler;
+import org.molgenis.data.index.job.IndexJobScheduler;
 import org.molgenis.data.meta.AttributeType;
 import org.molgenis.data.meta.MetaDataService;
 import org.molgenis.data.meta.model.*;
