+++ molgenis-platform-integration-tests/src/test/java/org/molgenis/integrationtest/platform/datatypeediting/FileTypeEditing.java
@@ -3,7 +3,7 @@ package org.molgenis.integrationtest.platform.datatypeediting;
 import org.molgenis.data.DataService;
 import org.molgenis.data.EntityManager;
 import org.molgenis.data.MolgenisDataException;
-import org.molgenis.data.elasticsearch.index.job.IndexJobScheduler;
+import org.molgenis.data.index.job.IndexJobScheduler;
 import org.molgenis.data.meta.MetaDataService;
 import org.molgenis.data.meta.model.AttributeFactory;
 import org.molgenis.data.meta.model.EntityType;
