+++ molgenis-platform-integration-tests/src/test/java/org/molgenis/integrationtest/platform/datatypeediting/OneToManyTypeEditing.java
@@ -2,7 +2,7 @@ package org.molgenis.integrationtest.platform.datatypeediting;
 
 import org.molgenis.data.DataService;
 import org.molgenis.data.EntityManager;
-import org.molgenis.data.elasticsearch.index.job.IndexJobScheduler;
+import org.molgenis.data.index.job.IndexJobScheduler;
 import org.molgenis.data.meta.MetaDataService;
 import org.molgenis.data.meta.model.AttributeFactory;
 import org.molgenis.data.meta.model.EntityType;
